+++ backend/baguni-api/src/main/java/baguni/api/application/pick/controller/PickApiController.java
@@ -65,6 +65,32 @@ public class PickApiController {
 						  .toList());
 	}
 
+	@GetMapping("/pagination")
+	@Operation(summary = "폴더 내 픽 페이지네이션 리스트 조회",
+		description = """
+				해당 폴더의 픽 리스트를 조회합니다.
+				커서 기반 페이지네이션 처리된 리스트가 반환됩니다.
+			""")
+	@ApiResponses(value = {
+		@ApiResponse(responseCode = "200", description = "픽 리스트 조회 성공")
+	})
+	public ResponseEntity<PickSliceResponse<PickApiResponse.Pick>> getFolderChildPickPagination(
+		@LoginUserId Long userId,
+		@Parameter(description = "조회할 폴더 ID", example = "1") @RequestParam Long folderId,
+		@Parameter(description = """
+				다음에 조회할 커서(cursor) 값입니다.
+				처음 페이지를 조회할 때는 0을 넣어주세요.
+				이후에는 응답으로 받은 lastCursor 값을 그대로 사용하시면 됩니다.
+				예시: lastCursor = 1 → 다음 요청에 cursor=1을 넣으면, 1은 제외하고 2부터 조회합니다.
+			""",
+			example = "0") @RequestParam(required = false, defaultValue = "0") Long cursor,
+		@Parameter(description = "한 페이지에 가져올 픽 개수", example = "20") @RequestParam(required = false, defaultValue = "20") int size
+	) {
+		var command = pickApiMapper.toReadPaginationCommand(userId, folderId, cursor, size);
+		var pickList = pickService.getFolderListChildPickPagination(command);
+		return ResponseEntity.ok(pickApiMapper.toSliceApiResponse(pickList));
+	}
+
 	@GetMapping("/search")
 	@Operation(summary = "픽 리스트 검색(페이지네이션)", description = "페이지네이션 처리 된 픽 리스트 검색")
 	@ApiResponses(value = {
